-put serialization utility functions into common header -add some FromChars tests when constexpr evaluation possible
